# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bindu API",
        "title": "Bindu API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/charts/visualization/{chart_id}": {
            "get": {
                "description": "Get chart by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chart"
                ],
                "summary": "Show a chart by id",
                "operationId": "get-chart-visualization-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chart ID",
                        "name": "chart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/charts/{chart_id}": {
            "get": {
                "description": "Get chart by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chart"
                ],
                "summary": "Show a chart by id",
                "operationId": "get-chart-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chart ID",
                        "name": "chart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Chart"
                        }
                    }
                }
            },
            "put": {
                "description": "Update chart by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chart"
                ],
                "summary": "Update a chart by id",
                "operationId": "update-chart-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chart ID",
                        "name": "chart_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Chart",
                        "name": "Chart",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/chart.chart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Chart"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete chart by ID",
                "tags": [
                    "Chart"
                ],
                "summary": "Delete a chart",
                "operationId": "delete-chart-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chart ID",
                        "name": "chart_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/permissions/organisations/{permission_id}": {
            "put": {
                "description": "Update Organisation permission by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Update a Organisation permission by id",
                "operationId": "update-org-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Body",
                        "name": "Permission",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/organisation.organisationPermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationPermission"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Organisation permission by ID",
                "tags": [
                    "Organisation_Permissions"
                ],
                "summary": "Delete a Organisation permission",
                "operationId": "delete-org-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/permissions/spaces/{permission_id}": {
            "get": {
                "description": "Get Space permissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Show a Space permissions",
                "operationId": "get-space-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SpacePermission"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Space permission by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Update a Space permission by id",
                "operationId": "update-space-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission Body",
                        "name": "Permission",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/space.spacePermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SpacePermission"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Space permission by ID",
                "tags": [
                    "Space_Permissions"
                ],
                "summary": "Delete a Space permission",
                "operationId": "delete-space-permission-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Permission ID",
                        "name": "permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "X-Space",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/spaces/{space_id}": {
            "put": {
                "description": "Update space",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space"
                ],
                "summary": "Update space",
                "operationId": "update-space",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Space Object",
                        "name": "Space",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.space"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Space"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete space",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Space"
                ],
                "summary": "Delete space",
                "operationId": "delete-space",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chart.chart": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "config": {
                    "type": "string"
                },
                "data_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "published_date": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "template_id": {
                    "type": "string"
                },
                "theme_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "charts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Chart"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_for_template": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.Chart": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Category"
                    }
                },
                "config": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "data_url": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "published_date": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "template": {
                    "$ref": "#/definitions/model.Template"
                },
                "template_id": {
                    "type": "string"
                },
                "theme": {
                    "$ref": "#/definitions/model.Theme"
                },
                "theme_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.OrganisationPermission": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Space": {
            "type": "object",
            "properties": {
                "contact_info": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fav_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "fav_icon_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_id": {
                    "type": "integer"
                },
                "logo_mobile": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_mobile_id": {
                    "type": "integer"
                },
                "mobile_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "mobile_icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "site_address": {
                    "type": "string"
                },
                "site_title": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "verification_codes": {
                    "type": "string"
                }
            }
        },
        "model.SpacePermission": {
            "type": "object",
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Theme": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "config": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
        "organisation.organisationPermission": {
            "type": "object",
            "required": [
                "organisation_id"
            ],
            "properties": {
                "organisation_id": {
                    "type": "integer"
                },
                "spaces": {
                    "type": "integer"
                }
            }
        },
...
        "space.space": {
            "type": "object",
            "required": [
                "name",
                "organisation_id"
            ],
            "properties": {
                "contact_info": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fav_icon_id": {
                    "type": "integer"
                },
                "logo_id": {
                    "type": "integer"
                },
                "logo_mobile_id": {
                    "type": "integer"
                },
                "mobile_icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "site_address": {
                    "type": "string"
                },
                "site_title": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "verification_codes": {
                    "type": "string"
                }
            }
        },
        "space.spacePermission": {
            "type": "object",
            "required": [
                "space_id"
            ],
            "properties": {
                "charts": {
                    "type": "integer"
                },
                "space_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}